refactor(buildLogic/convention/plugins/codegen/*): made internal; changed package; no clean- wizard
