+++ build-logic/convention/src/main/kotlin/corp/tbm/cleanarchitecturemapper/buildLogic/convention/plugins/codegen/VisitorPlugin.kt
@@ -7,7 +7,7 @@ import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.dependencies
 
-class VisitorPlugin : Plugin<Project> {
+internal class VisitorPlugin : Plugin<Project> {
 
     override fun apply(target: Project) {
         with(target) {
